requirements.txt
#Environment snapshot created from pip list
#Use 'pip install -r requirements.txt' to replicate this environment
#Core Web and UI Frameworks
flask==3.0.3
dash==2.17.0
dash-bio==1.2.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Werkzeug==3.0.3
click==8.1.7
blinker==1.8.2
MarkupSafe==2.1.5

#Bioinformatics & Scientific Computing
viennarna==2.7.0
parmed==4.3.0
biopython==1.84
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

#Plotting Libraries
plotly==5.22.0
matplotlib==3.9.0
seaborn==0.13.2

#Data Handling & Excel
openpyxl==3.1.4
et-xmlfile==1.1.0
xlsxwriter==3.2.0
tabulate==0.9.0
xlrd==2.0.1

#HDF5 / C-Trap / OT
h5py==3.11.0
lumicks.pylake==1.6.1

#Image Processing (Optional)
imageio==2.34.1
scikit-image==0.24.0
tifffile==2024.5.22

#Machine Learning & Utilities (Optional)
scikit-learn==1.5.0
joblib==1.4.2

#Other Helper Libraries
tqdm==4.66.4
python-dateutil==2.9.0.post0
requests==2.32.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Babel==2.15.0
Brotli==1.1.0
build==1.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
exceptiongroup==1.2.1
fastjsonschema==2.20.0
fonttools==4.53.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.2.1
iniconfig==2.0.0
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
kiwisolver==1.4.5
lazy-loader==0.4
lxml==5.2.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
packaging==24.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
prometheus-client==0.20.0
prompt-toolkit==3.0.42
psutil==5.9.8
pyarrow==16.1.0
pybind11==2.13.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-json-logger==2.0.7
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
retrying==1.3.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.18.6
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
tenacity==8.4.1
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
typing-extensions==4.12.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.1